Renames to `prevent-comment-loss`
